Adding comments and "matches.py" endpoind review

+++ app/routers/matches.py
@@ -10,8 +10,12 @@ router = APIRouter()
     # GET Match method
 @router.get("/matches/{match_id}", response_model=MatchPublic)
 def get_match(current_player: ActivePlayer, session: SessionDep, match_id: int):
-    return session.get(Match, match_id)
-
+    match = session.get(Match, match_id)
+    if match:
+        return match
+    else:
+        raise HTTPException(status_code=404, detail="Unable to find match with mentioned id.")
+    
     # GET Matches list
 @router.get("/matches/", response_model=list[MatchPublic])
 def get_match_list(current_player: ActivePlayer, session: SessionDep, offset: int = 0, limit: int = Query(default=100, le=100)) -> list[Match]:
@@ -36,7 +40,7 @@ def post_match(current_player: ActivePlayer, match: MatchCreate, session: Sessio
 def update_match_settings(current_player: ActivePlayer, match_id: int, match: MatchUpdateSettings, session: SessionDep):
     match_db = session.get(Match, match_id)
     if not match_db:
-        raise HTTPException(status_code=404, detail="Hero not found")
+        raise HTTPException(status_code=404, detail="Match not found")
     match_data = match.model_dump(exclude_unset=True)
     match_db.sqlmodel_update(match_data)
     session.add(match_db)
@@ -49,6 +53,9 @@ def update_match_settings(current_player: ActivePlayer, match_id: int, match: Ma
 @router.delete("/matches/{match_id}")
 def delete_match(current_player: ActivePlayer, match_id: int, session: SessionDep):
     db_match = session.get(Match, match_id)
-    session.delete(db_match)
-    session.commit()
-    return {"ok": True}
+    if db_match:
+        session.delete(db_match)
+        session.commit()
+        return {"ok": True}
+    else:
+        raise HTTPException(status_code=404, detail="Unable to find match with mentioned id.")
